requirements.txt
transformers>=4.34.0
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.2.0
datasets>=2.14.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.66.0
python-dotenv>=1.0.0
accelerate>=0.23.0
bitsandbytes>=0.41.0
sentencepiece>=0.1.99
huggingface-hub>=0.17.0 